+++ lib/rubocop/cop/style/if_unless_modifier.rb
@@ -9,6 +9,8 @@ module RuboCop
       class IfUnlessModifier < Cop
         include StatementModifier
 
+        ASSIGNMENT_TYPES = [:lvasgn, :casgn, :cvasgn, :gvasgn, :ivasgn, :masgn]
+
         def message(keyword)
           "Favor modifier `#{keyword}` usage when having a single-line body." \
           ' Another good alternative is the usage of control flow `&&`/`||`.'
@@ -20,15 +22,36 @@ module RuboCop
           return if modifier_if?(node)
           return if elsif?(node)
           return if if_else?(node)
-          # Accept cases that require parentheses around modifier if statement.
           return if chained?(node)
           return unless fit_within_line_as_modifier_form?(node)
           add_offense(node, :keyword, message(node.loc.keyword.source))
         end
 
         def chained?(node)
-          ancestor = node.ancestors.first
-          ancestor && ancestor.send_type?
+          # Don't register offense for `if ... end.method`
+          return false if node.parent.nil? || !node.parent.send_type?
+          receiver = node.parent.children[0]
+          node.equal?(receiver)
+        end
+
+        def parenthesize?(node)
+          # Parenthesize corrected expression if changing to modifier-if form
+          # would change the meaning of the parent expression
+          # (due to the low operator precedence of modifier-if)
+          return false if node.parent.nil?
+          return true if ASSIGNMENT_TYPES.include?(node.parent.type)
+
+          if node.parent.send_type?
+            _receiver, _name, *args = *node.parent
+            return !method_uses_parens?(node.parent, args.first)
+          end
+
+          false
+        end
+
+        def method_uses_parens?(node, limit)
+          source = node.loc.expression.source_line[0...limit.loc.column]
+          source =~ /\s*\(\s*$/
         end
 
         def autocorrect(node)
@@ -43,6 +66,7 @@ module RuboCop
           if first_line_comment
             oneline << ' ' << first_line_comment.loc.expression.source
           end
+          oneline = "(#{oneline})" if parenthesize?(node)
 
           ->(corrector) { corrector.replace(node.loc.expression, oneline) }
         end